lectures/7.context/chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "asif-ml-developer", email = "asif.ml.developer@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.1",
    "google-generativeai>=0.8.4",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]

[project.scripts]
chatbot = "chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
